Respect --namespace filter without setting dummy --namespace-regex
Currently, if your project has two test namespaces, `a-test` and
`b-test` if you call `clj -m test.runner -n b-test`, the tests from both
namespaces will run. This is because the default ns regex `#"*-test"`
allows both namespaces.

One fix for this is to change the command line: `clj -m test.runner -n
b-test -r zzz`. But that feels like a hack. This commit prevents the
need for the hack.

One criticism of this fix is that it ignores `-r` whenever `-n` is set.
To me, this makes sense, but if you would expect `-r` and `-n` to apply
simultaneously, let me know how you would like them to interact.